requirements.txt
openai>=1.0.0
python-dotenv
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
redis>=5.0.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4